Enable one test to be run at once.

+++ scripts/test.sh
@@ -1,15 +1,17 @@
 #!/bin/sh
 
-TESTS=("blog" "gallery" "index")
+TESTS=("blog" "gallery" "index" "directories")
 
 for TEST in ${TESTS[*]}; do
-    echo "#"
-    echo "# Running test: $TEST"
-    echo "#"
-    cd "tests/$TEST"
-    rm -rf build/
-    luapress init "/tests/$TEST/index"
-    luapress
-    cd ../..
-    echo
+    if [[ "$1" = "" ]] || [[ "$1" = "$TEST" ]]; then
+        echo "#"
+        echo "# Running test: $TEST"
+        echo "#"
+        cd "tests/$TEST"
+        rm -rf build/
+        luapress init "/tests/$TEST/index"
+        luapress
+        cd ../..
+        echo
+    fi
 done
